refactor: 패키지 계층 통일
- DTO 패키지를 Serive내로 이동
