+++ backend/src/main/java/com/woowacourse/momo/member/service/dto/request/ChangePasswordRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.member.dto.request;
+package com.woowacourse.momo.member.service.dto.request;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
