+++ backend/src/main/java/com/woowacourse/momo/participant/controller/ParticipantController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.auth.config.AuthenticationPrincipal;
-import com.woowacourse.momo.member.dto.response.MemberResponse;
+import com.woowacourse.momo.member.service.dto.response.MemberResponse;
 import com.woowacourse.momo.participant.service.ParticipantService;
 
 @RequiredArgsConstructor
