+++ backend/src/main/java/com/woowacourse/momo/participant/service/ParticipantService.java
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 import com.woowacourse.momo.group.domain.group.Group;
 import com.woowacourse.momo.group.service.GroupFindService;
 import com.woowacourse.momo.member.domain.Member;
-import com.woowacourse.momo.member.dto.response.MemberResponse;
-import com.woowacourse.momo.member.dto.response.MemberResponseAssembler;
+import com.woowacourse.momo.member.service.dto.response.MemberResponse;
+import com.woowacourse.momo.member.service.dto.response.MemberResponseAssembler;
 import com.woowacourse.momo.member.service.MemberFindService;
 
 @RequiredArgsConstructor
